split out apic module
