+++ kernel/CMakeLists.txt
@@ -12,6 +12,7 @@ target_sources(kernel PRIVATE
     "${CMAKE_CURRENT_SOURCE_DIR}/serial.c"
     "${CMAKE_CURRENT_SOURCE_DIR}/processors.c"
     "${CMAKE_CURRENT_SOURCE_DIR}/interrupts.c"
+    "${CMAKE_CURRENT_SOURCE_DIR}/apic.c"
     "${CMAKE_CURRENT_SOURCE_DIR}/x86_64.asm"
 )
